Change practitioner height from integer to float

+++ db/migrate/20230117001842_add_height_to_practitioner_profiles.rb
@@ -1,5 +1,5 @@
 class AddHeightToPractitionerProfiles < ActiveRecord::Migration[7.0]
   def change
-    add_column :practitioner_profiles, :height, :integer
+    add_column :practitioner_profiles, :height, :float
   end
 end
